Ignore Non-Root JARs
Previously, when detecting the single JAR that indicated that an archive
should be expanded, the algorithm would descend and inspect all children
regardless of how deep they were.  Unfortunately this often detected "wrapper"
application JARs like the Maven wrapper and would expand it.

This change ensures that only a single archive, located in the root of the
application folder would be a viable candidate.

Signed-off-by: Ben Hale <bhale@pivotal.io>